Create Philatelie.java
Philately
A philatelist wishes to estimate at which price he could sell his stamps. The
purpose of this exercise is to write a program which allows him to do so.
The provided code does the following :
• creates a collection of stamps (rare or commemorative) ;
• and displays the price of every stamp of this collection.
The hierarchy of Timbre (Stamp in french) is missing and it is what we ask you
to write.
A stamp is characterized by : its code, its année d’émission (issue year), its pays
d’origine (country of origin) and its valeur faciale (denomination value) in francs
(the equivalent in francs of the value printed on the stamp).
Our philatelist distinguishes two broad categories of stamps, which are distinguished by the way the selling price is computed.
• the rare stamps (class Rare) : possesses an extra attribute indicating the
number of copies that are inventoried throughout the world ;
• the commemorative stamps (Commemoratif) : without any particular
specific attribute.
Selling price of the stamps The selling price of a timbre (stamp) of any kind
is its denomination value, if the stamp is less than 5 years old. If not, the selling
price is the denomination value multiplied by the age of the stamp and by the
coefficient 2.5.
The selling price of a rare stamp starts from a prix de base (base price) : 600 francs
if the number of the inventoried copies is less than 100, 400 francs if the number of
copies is between 100 and 1000 and 50 francs in any other case. The selling price
of a rare stamp is then given by the formula prix_base * (age_timbre /
10.0). The different constants involved in the calculation are given in the
provided file.
The selling price of a Commemoratif stamp is twice as much as the price of a
generic one.
Basic stamp The public methos of the class Timbre are :
• constructors that conform to the provided main() method, with the following order of the parameters : the code, the issue year, the country of
origin, and the denomination value ; each one of this parameters takes a
default value : VALEUR_TIMBRE_DEFAUT for the denomination value,
PAYS_DEFAUT for the country, ANNEE_COURANTE for the year and CODE_DEFAUT
for the code ; the construction of a stamp can be thus done with zero, one,
two, three or four parameters (respecting the order given above) ;
• a method vente() returning the selling price in the format of a double ;
• a method toString() producing a representation of a Timbre respecting strictly the following format :
Timbre de code <code> datant de <annee> (provenance <pays>)
ayant pour valeur faciale <valeur faciale> francs
in only one line where <code> is to be replaced by the code of the stamp,
<annee> by its issue year, <pays>, by its country of origin and <valeur faciale>,
by its denomination value ;
the method age() returns the age of a stamp in the format of an int
(difference between ANNEE_COURANTE (current year) and the issue year
of the stamp) ;
• the getters getCode(), getAnnee(), getPays() and getValeurFaciale().
Rare stamps The class Rare has to provide as public methods :
constructors that conform to the provided main(), with the following order
for the parameters : the code, the issue year, the country of origin, the
denomination value and the number of inventoried copies. The number of
copies should always be provided when the constructors are called. The
other parameters, if they are not provided, have the default values that are
indicated for the basic stamps ;
• a getter getExemplaires() ;
• a method toString() producing a representation of a Rare respecting
strictly the following format :
Timbre de code <code> datant de <annee> (provenance <pays>)
ayant pour valeur faciale <valeur faciale> francs
in only one line then an End Of Line followed by
Nombre d’exemplaires -> <exemplaires>
where <code> is to be replaced by the code of the stamp, <annee> by its
issue year, <pays> by its country of origin and <valeur faciale>,
by its denomination value and <exemplaires> by the number of inventoried copies ;
Commemorative stamps The class Commemoratif has to provide as public
methods :
• constructors that conform to the provided main(). Its parameters, if they
are not provided, they have the default values that are indicated for the basic
stamps ;
• a method toString() producing a representation of a Commemoratif
respecting strictly the following format :
Timbre de code <code> datant de <annee> (provenance <pays>)
ayant pour valeur faciale <valeur faciale> francs
in only one line then an End Of Line followed by
Timbre celebrant un evenement
where <code> is to be replaced by the code of the stamp, <annee> by its
issue year, <pays> by its country of origin and <valeur faciale>,
by its denomination value.
It should be possible to calculate the selling price of a stamp Commemoratif
using a method double vente().
You are therefore asked to code the hierarchy of classes according to this description. Avoid code duplication, pay attention to the access modifiers of the attributes
and name the classes as suggested in the statement.
By the way, you will assume that the classes Rare and Commemoratif do not
inherit the one from the other.
1.2 Execution example
Timbre de code Guarana-4574 datant de 1960 (provenance Mexique) ayant pour valeur faciale 0.2 francs
Nombre d’exemplaires -> 98
Prix vente : 3360.0 francs
Timbre de code 700eme-501 datant de 2002 (provenance Suisse) ayant pour valeur faciale 1.5 francs
Timbre celebrant un evenement
Prix vente : 105.0 francs
Timbre de code Setchuan-302 datant de 2004 (provenance Chine) ayant pour valeur faciale 0.2 francs
Prix vente : 6.000000000000001 francs
Timbre de code Yoddle-201 datant de 1916 (provenance Suisse) ayant pour valeur faciale 0.8 francs
Nombre d’exemplaires -> 3
Prix vente : 6000.0 francs